# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmAid API
  description: Disaster relief coordination platform connecting surplus resources with needs
  version: 1.0.0
  contact:
    name: SwarmAid Team
    email: api@swarmaid.org
servers:
  - url: https://api.swarmaid.org
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MatchRecommendation:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        demandId:
          type: string
        supplierId:
          type: string
        recipientId:
          type: string
        score:
          type: number
        scoreBreakdown:
          type: object
        distanceMiles:
          type: number
        estimatedDurationMinutes:
          type: number
        status:
          type: string
          enum: [pending, accepted, rejected, scheduled, in_progress, completed, failed]
        routePlan:
          type: object
        complianceStatus:
          type: string
          enum: [pending, passed, blocked, override]
...
paths:
...
  /matches:
    get:
      summary: List match recommendations
      tags: [Matches]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRecommendation'
  /matches/{id}:
    get:
      summary: Get match by ID
      tags: [Matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRecommendation'
  /matches/{id}/accept:
    post:
      summary: Accept match recommendation
      tags: [Matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRecommendation'
...